+++ NEWS.md
@@ -1,3 +1,9 @@
+enmSdmX 1.0.2
+===========
+o Fixed issue in nearestGeogPoints() when polygon lay under centroid
+o Fixed bug with "table" call of getCRS()
+o getCRS() (no arguments) now displays shiny table of all available CRSs
+
 enmSdmX 1.0.1
 ===========
 o Fixed bug experienced by some users using predictEnmSdm() and predictMaxNet() (Thank you, Nikki!)